feat(tiller): sort list in tiller, not helm
