+++ cmd/tiller/release_server.go
@@ -5,6 +5,7 @@ import (
 	"errors"
 	"fmt"
 	"log"
+	"sort"
 
 	"github.com/kubernetes/helm/cmd/tiller/environment"
 	"github.com/kubernetes/helm/pkg/proto/hapi/release"
@@ -48,6 +49,17 @@ func (s *releaseServer) ListReleases(req *services.ListReleasesRequest, stream s
 	}
 
 	total := int64(len(rels))
+
+	if req.SortBy != services.ListSort_UNKNOWN {
+		log.Printf("Sorting by %q", req.SortBy)
+	}
+
+	switch req.SortBy {
+	case services.ListSort_NAME:
+		sort.Sort(byName(rels))
+	case services.ListSort_LAST_RELEASED:
+		sort.Sort(byDate(rels))
+	}
 
 	l := int64(len(rels))
 	if req.Offset > 0 {
@@ -223,4 +235,27 @@ func (s *releaseServer) UninstallRelease(c ctx.Context, req *services.UninstallR
 
 	res := services.UninstallReleaseResponse{Release: rel}
 	return &res, nil
+}
+
+// byName implements the sort.Interface for []*release.Release.
+type byName []*release.Release
+
+func (r byName) Len() int {
+	return len(r)
+}
+func (r byName) Swap(p, q int) {
+	r[p], r[q] = r[q], r[p]
+}
+func (r byName) Less(i, j int) bool {
+	return r[i].Name < r[j].Name
+}
+
+type byDate []*release.Release
+
+func (r byDate) Len() int { return len(r) }
+func (r byDate) Swap(p, q int) {
+	r[p], r[q] = r[q], r[p]
+}
+func (r byDate) Less(p, q int) bool {
+	return r[p].Info.LastDeployed.Seconds < r[q].Info.LastDeployed.Seconds
 }
